Make intermediary structures private again
