+++ pkg/services/provisioning/datasources/configreader/disk_config_reader_test.go
@@ -5,12 +5,12 @@ import (
 	"os"
 	"testing"
 
+	. "github.com/smartystreets/goconvey/convey"
+
 	"github.com/grafana/grafana/pkg/bus"
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/provisioning/datasources"
-
-	. "github.com/smartystreets/goconvey/convey"
 )
 
 var (
