+++ pkg/services/provisioning/datasources/configreader/vcs_config_reader.go
@@ -27,17 +27,17 @@ func (cr *vcsConfigReader) ReadConfigs(ctx context.Context) ([]*datasources.Conf
 	}
 
 	for _, versionedDc := range datasourcesMap {
-		cfg := datasources.ConfigsV1{Log: cr.log}
+		cfg := configsV1{Log: cr.log}
 		cfg.APIVersion = 1
 
-		dcCfg := datasources.UpsertDataSourceFromConfigV1{}
+		dcCfg := upsertDataSourceFromConfigV1{}
 		err := json.Unmarshal(versionedDc.Data, &dcCfg)
 		if err != nil {
 			return nil, err
 		}
 
 		cfg.Datasources = append(cfg.Datasources, &dcCfg)
-		configs = append(configs, cfg.MapToDatasourceFromConfig(1))
+		configs = append(configs, cfg.mapToDatasourceFromConfig(1))
 	}
 
 	return configs, nil
