+++ pkg/services/provisioning/datasources/configreader/vcs_config_reader_test.go
@@ -4,12 +4,13 @@ import (
 	"context"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/services/provisioning/datasources"
 	"github.com/grafana/grafana/pkg/services/vcs"
 	"github.com/grafana/grafana/pkg/services/vcs/vcsmock"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 // TODO what do we do with typeName and readOnly (expected editable)
